django_framework/django_mvt/requirements.txt
django-extensions
django-ckeditor # 富文本编辑器
django-summernote # 富文本编辑器

djangorestframework